coll: remove unnecessary SMP CVARs
These CVARs were originally added to let the user disable SMP
collectives in case there are bugs or performance problems.  The code
has been used in production for a while now, so the risk of bugs is
lesser now.  And performance aspects need to be handled by tuning what
cases a particular algorithm is enabled in, instead of simply
disabling these algorithms.