mtest/f77: Fix wrong assumption on MPI_Test
In ext/c2f2c.c, c2frequest_ was assuming that MPI_Test would always
return flag==true (request completed). However, this is not guaranteed
especially when the MPI library implements an asynchronous progress.

This patch adds a check on the flag returned by MPI_Test to make sure
MPI_Test_cancelled is called only when the request is complete.