config: add maint/config_filter.pl
This script is to be run by AC_CONFIG_COMMANDS in mpich's configure.ac,
after all the AC_OUTPUT files are generated -- in particular, mpi.h and
mpichconf.h.

Configure script's function is to gather user options and do system feature
detections. That includes checking the sizes of language datatypes.  This
script does further processing based on the results in mpichconf.h. It does
following tasks:

* Putting in the size into MPI datatype handles in mpi.h.
* typedef MPI_Aint, MPI_Offset, MPI_Count in mpi.h.
* Putting in handle decimal values in mpif.h.
* Define following convenience macros in mpi_ext.h
    Define HAVE_datatype macros for all types that have non-zero sizes.
    Define MPIR_xxx_CTYPE for FORTRAN and CXX datatypes.
    Define MPIR_xxx_FMT_[dx] and MPIR_xxx_MAX for MPI_Aint, MPI_Offset, and MPI_Count.