config: add config_filter.pl to AC_CONFIG_COMMANDS
Remove the command to do fortran "!" and "REAL*8" substitution with
maint/config_filter.pl, that does the substitution and more.

TODO: we should define $has_exclaim and $has_fort_real8 as macros in
mpichconf.h, where config_filter.pl picks up the rest of the config
flags, and remove the environment export.