testsuite: longer time limit for some rma lockall tests
Currently rma/lockall_dt_ tests with non-contiguous datatype are very
slow for ofi direct-nm configuration, potentially will be improved after
the a datatype refactor. Since these tests are running on the
border-line of the default 3 min. timeLimit, this patch increase the
limit to 10 minutes. It should prevent triggering test failures while we
monitor the situations.