mpl: add unit_test facility
MPL is a good candidate for unit testing, where individual
funcationality can be tested independent of the rest of the code.

This commit adds unit_test.pl, which looks for .c files inside the
"test" folder. Each .c file is to be compiled and run individually and
its return code is used to mark OK or FAIL.

The script is general enough and can be potentially used for unit-tests
outside MPL.