ch4: remove MPIDIU_avt_release_ref in MPID_Finalize
The additional entries in MPIDI_global.av_table_list are allocated with
dynamic processes, and the reference counting (tied to communicator
creation and free) should have freed all of them at finalize stage.

The init-time node_map and av_table are separately allocated and freed.

If this hypothesis is wrong, the memtracing in debug build should catch
it (in spawn tests).