init: refactor init_thread_cs
During init we only need MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX and we need
use it as soon as we can. This commit separates the "allfunc" mutex from
the rest of the mutexes so the rest of the configuration dependant
mutexes can be easily further refactored.

We also replaced the error handling from MPL_thread_init with
MPL_Assert. In general, we should use assert for unexpected
configuration/build or coding errors; while return mpi_errno for run-time
errors. Thread initialization is of the former type.