ch4/ofi: Avoid a race condition for MT probe with shared vci
Changes are introduced to ensure that a thread waiting during
a probe call will unblock only after all the related data have
been copied to the request object during progress.

This fixes an intermittent race condition that can occur under
multi-threaded probe when the threads are sharing a vci. Race can
occur if an OFI peek request submitted by a thread is processed
by a different thread during progress, which is possible when
the threads share a vci.

Note: Part of this fix was merged with an earlier patch:
      bdc184e8fe1a262bf13137c61aed7b7326607a3d