coll: implement custom MPIR_Barrier_intra_fallback
It is a duplicate of MPIR_Barrier_intra_dissemination except use
finite-sized message rather than zero sized NULL buffer. This is needed
as the OFI max_buffered_size is of size_t and a zero sized message will
always be sent with fi_inject even when max_buffered_size set to 0.

While this can be more easily solved by changing the max_buffered_size
type to signed type, such as MPI_Aint, and setting max_buffered_size to
-1, for the consistency, it may require such type changes on a wholesale
scale. Reference PR #3826. On the otherhand, we likely need a
independent fallback rountine for barrier any way, as the fallback code
emphasize on dependency and stability, the run-time code will focus on
performance.