ch4/coll: Fix error detection and propagation in Bcast
In alpha composition for Bcast, root of Bcast and rank 0 of that node
does a send-recv. After the recv, the amount of data should be matched
to the expected value. In case of a mismatch, appropriate error flags
should be set and correctly propagated further. This data mismatch
checking happens only when "error checking" is enabled in the library.

Conflicts:
	src/mpid/ch4/src/ch4_coll_impl.h