ch4: provide error buffer in fi_cq_err_entry
This way, the error buffer is private to each thread. If we
don't pass the buffer during fi_cq_readerr call, OFI will copy
its internal buffer address and return it to the threads. As a
result, the threads can share an error buffer.

This change makes the call to fi_cq_readerr thread-safe, which
is important for the lockless threading model.