ch4/ofi: Correctly measure size when using probe
Previously, OFI returned the size of the last chunk when proving a large
message. Instead, it should have been using the size of the entire
message that is transfered in the control message. To accomplish this,
this patch adds another protocol bit to signal a large message. When the
user calls a probe function, the OFI netmod looks for a matching control
message. If such a message doesn't exist yet, a temporary request is put
into the huge message queue, which is removed later when the control
message arrives.