shm: remove alloc/commit shmem allocation mechanism
Currently, allocating shared memory involves requesting memory through
invocations of MPIDU_shm_seg_alloc followed by one final call to
MPIDU_shm_seg_commit. This mechanism was put in place to overload the
segment with additional space, for example, for share memory barrier.
This mechanism is no longer needed and we move to a more intuitive
interface in which memory is allocated and free directly.