ofi: consolidate macros to MPIDI_OFI_ENABLE_ENDPOINT_BITS
We only want to enable ep_idx when SHARED_ENDPOINT is enabled. However,
when MPIDI_OFI_ENABLE_RUNTIME_CHECKS is enabled,
MPIDI_OFI_ENABLE_SCALABLE_ENDPOINTS is defined to a runtime variable,
therefore, can't be used in preprocessor. To work around, ep_idx is
always enabled for that case. The previous double macro switch is
difficult to read. This commit consolidates the pattern to use a single
macro -- MPIDI_OFI_ENABLE_ENDPOINT_BITS