ofi: add vci_lock to protect the progress in ofi_spawn
The normal progress is protected with vci_lock, the progress in
ofi_spawn is unprotected and it needs to be protected.  We
customized the OFI PROGRESS macros here to add the vci_locks as
a work-around.

A proper fix is to restructure the locks into finer grain and remove the
yielding.