ch4: add fields for preallocated per-comm recv request.
The preallocated recv request aims to reduce the overhead of request
allocation in simple irecv operation. The preallocated request must be
allocated per-comm (if info hint set for the comm) because it need to
track the context id. Two requests are need to to track the
communication using different conext offsets (for PT2PT and COLL). When
the info hint is not set (default), no request is preallocated.