ch3: use SIZEOF_VOID_P instead of SIZEOF_OPA_PTR_T
SIZEOF_OPA_PTR_T is not defined. It probably once defined but got
refactored away at some point. To support it will be quite messy as the
configure don't really decide which opa atomic primitive it is being
used -- opa_primitives.h does.

For practical purpose, this is usually the same as SIZEOF_VOID_P. And
since we are going to replace openpa with mpl atomics, let's take the
cheap fix for now. We need ensure the size of MPL_Atomic_ptr_t is the
same as void * in the future.