mpl: add no_instrument_function attribute to avoid recursive calls
The logging call calls some function that need be excluded from tracing
or it will result in dead recursive calls.

Some of the timer functions are not used by the logging calls, but since
they could be used by sub-calls and in general, we won't desire to log
timing calls anyway, so they are also annotated.