handlemem: cleaner indirect type declaration
The existing code declares the indirect pointer of type
`*(*indirect)[]`. This is difficult to read and some compiler complains
about it, such as Oracle Developer Studion (suncc).

Indirect is a dynamically allocated array of void pointers, each points
to a block of objects. Therefore, the type should simply be void **.