config: remove the check for invalid prototype
In PAC_C_CHECK_COMPILER_OPTION, after we check for whether a cflag works
with a trivial program, we also check if it works with a program that
contains a warning. First, the check won't work if the presumed warning
is not triggered. In our choses code, the use of uninitialized variable
warning does not trigger in icc by default. Second, if it gives warning,
then it won't work with -Werror or its like. Sometime we want to check
and add -Werror, and its supposed behavior is very unclear with this
check. Third, the original issue seems to be a particular odd behavior
of a very old gcc (gcc 4.2). Therefore, we are removing this check.