mpir/topo: rename GID_INVALID to GID_ROOT
We rename MPIR_HWTOPO_GID_INVALID to MPIR_HWTOPO_GID_ROOT and assign the
machine gid to it. Returning the machine gid allows to always have a
valid gid returned by the topology layer to the caller, even for cases
in which the query does not find any object in the topology tree (e.g.,
if we query for infiniband card and none is present in the node). One
example is comm_split_type. In this case since we use the gid as color
for the communicator split we always want every process to get a valid
gid.