config: remove AC_USE_SYSTEM_EXTENSIONS
AC_USE_SYSTEM_EXTENSIONS will define:
_GNU_SOURCE
__EXTENSIONS__
_POSIX_PTHREAD_SEMANTICS
_TANDEM_SOURCE
_ALL_SOURCE
_POSIX_1_SOURCE
_MINIX

i.e. It will enable extensions even it is not defined by POSIX.

This is against our other policy such as using conservative C standards
to ensure portability. In addition, it is messy to multiple packages to
duplicate these definitions or it will bring inconsistency during
configure. Therefore, it is better off to not enable them and solve the
protability inside MPL.