ch4: add _GNU_SOURCE due to the usage of MAP_ANON
Since we removed AC_USE_SYSTEM_EXTENSIONS, specific linux only features,
such as MAP_ANON, need be individually enabled with _GNU_SOURCE.

The code will break on non-supported platform such as BSD. A better
solution is to have configure option and guard the code with it. Refer
to github issue #4173.