shm: we have to assert if attempt freeing an invalid buffer
MPIDU_Init_shm_free looks up the passed pointer into a linked list to
find the corresponding memory handle, created when memory was allocated
with MPIDU_Init_shm_alloc. If the pointer does not have any entry in the
linked list (e.g., because the user passes a buffer that was allocated
using another allocation mechanism, such as malloc) we assert.