romio: add long long cast to silence printf format warnings
On 64-bit machines, long has the same size as long long. MPICH configures
MPI_OFFSET to long, but ROMIO will configure it to long long. This
raises warnings when we debug print it with the format "%lld" (or %ld).
This commit patches the warnings by casting the argument to "long long".

Many parts of ROMIO code already uses this kind of cast to silence
warnings. This partch merely extends the practice to the rest of code.

I do not enjoy this kind of patching but lack of better alternatives.
Ideally, compilers should offer "auto" format such as `%v` for at least
integer and floating types.