testsuite: remove util:all target from testlist
Now that util folder contains code that may not be needed, the all
target may not be able to build. For example, mtest_f77.c only can be
built when fortran is enabled.

The makefiles should build the mtest_xxx.o object as needed, so
pre-build these utility objects is unnecessary.