ch4/ofi: Use FI_ORDER_ATOMIC_ instead of FI_ORDER_
In OFI 1.8, more relaxed ordering requirements were introduced to allow
a provider to only guarantee a specific ordering level for atomics,
rather than for all messages. As most hardware-offloaded networks are
likely to use these flags, and MPICH doesn't require the ordering for
non-atomic operations, this commit changes all uses of `FI_ORDER_*` to
be `FI_ORDER_ATOMIC_*`.

Because these new flags didn't exist before OFI 1.8, there is a compile
time check to look for the correct version of these flags to be used.
For older versions of libfabric, the older `FI_ORDER_*` flags will be
used instead. Of course, this means that compiling with an older version
of libfabric will not work when running with a newer version of
libfabric (or at least will result in unexpected behavior).