ch4: always use workq when it is configured
This commit decouples the thread-safety of workq from that of regular
communications. The decoupling allows refactoring of per-vci thread
safety to proceed. One of change we are considering is to move
thread-safety of send/recv operation down to netmod/shmmod layer.
Without the decoupling, the workq code need to be moved down too,
making things complicated.