ch4/ofi: register ofi internal am handler id in ch4_types.h
The way ofi hacks an id based on MPIDI_AM_HANDLERS_MAX is very ugly.
There is no way to prevent other modules don't cllide with it.

For now we simply inserts the static handler id the global table in
ch4_types.h. Left as TODO, if we wish to keep sub-modules handler id
hidden, we could use the dynamic registration and store the returned
handler id.