mpl: Move MPL_ilog function into tree_utils
The function to calculate log with integer base is not implemented
generic enough. It produces different results from one would expect.
Seems like it is specific to tree based calculations. So moving it from
MPL to tree utils, which will avoid any incorrect results in future, in
case this function is used somewhere else.