ch4: use MPI_Aint for count parameter in ch4_recv
This fixes a bug that is hiding for so long.

We had integer overflow due to MPI_Aint to int, become a negative.
Typically negative will cause severe errors and trigger test failures.
However, our code uses size_t a lot and we cast to (size_t) often.
The conversion converts the count to a very large number but no-longer
negative, and tests failed to trigger due Recv is allowed having bigger
buffer.

This again highlights my point that we should banish size_t (and all
unsigned type) whenever number semantics is involved (such as count).
With numbers, even when they are constants, we do arithmetics,
comparisons and conversions. With unsigned number sitting right at
under-flow edge (0), it is simply too dangerous to use.