ch4/am: add routine MPIDIG_get_recv_data
Some special am transport, OFI to be specific, can't directly call
MPIDIG_recv_copy nor MPIDIG_recv_copy_seg as it need do rdma-read
instead of direct memcpy. IOV is still needed in this case. Rather than
exposing the data structure that is semantically not needed by the
transport, this commit adds a convenience function within mpidig_msg.h
to get iov on demand.

The goal is to remove the iov from the general callback interface as
they are cumbersome and most transport do not need.