izem: remove izem-sync and izem-atomic related code.
izem-sync is for locks and mutexes, but it internally does polling.
If we ever have thread oversubscription, this degrades performance to
an extent where all of our threading tests timeout.  This is too risky
to enable in a production release.

izem-atomics are not used in mpich anymore because we use MPL atomics.