thread: remove nonrecursive mutex calls
Nonrecursive mutex calls were only used for POBJ code.  First, this is
incorrect and assumes that ROMIO (which is the only user of recursive
locks) will always be disabled.  Second, the cost of checking for
recursion is a TLS lookup for the thread ID, which, on modern
compilers, is negligible compared with the mutex acquisition (which is
at least an atomic operation in the best case).