ch3/nemesis: always force lock-free queues.
The non-lock-free queue code is completely untested at this point.  It
was added when nemesis relied on x86-specific atomic operations for
lock-free queues.  Now we rely on MPL, which internally has lock-based
atomics for platforms that do not have native atomics (if those even
exist), so this is no longer needed.