ch4/ofi: remove call to MPIDI_OFI_do_emulated_inject
The MPIDI_OFI_do_inject is used in MPIDI_NM_am_send_hdr and is used for
am protocols that does not have a payload. They usually fit inside the
eager limit. For example, for current posix, it won't work if the header
does not fit inside eager limit. This commit removes the emulated inject
fallback and replaced with an assertion. For prtocols that have very
large headers can simply call the regular MPIDI_NM_am_isend with zero
payload.