thread: remove MPICH TLS requirements
MPICH TLS is only used to support a thread-safe MPIR_Strerror.  This
is unnecessary and is inconsistent with other POSIX functions such as
strerror_r.  Instead, follow the same model as what strerror_r does in
POSIX 2001, i.e., allow the caller to provide the buffer to place the
error string in.  Since the buffer is a stack variable, we do not have
to worry about thread safety or cleaning up of the buffer when the
thread terminates.