hydra: allow PMI_PORT to be used
PMI_FD uses UNIX sockets through a pipe.  Unfortunately, when passed
through debuggers such as lldb and idb, the UNIX socket pipes seem to
be closed, so the MPI process cannot talk to the proxy.  This patch
allows the MPI processes to connect back to the proxy through a TCP
socket, instead of a UNIX socket.  This model is not always reliable
because it assumes that the network configuration for the local node
is correctly setup, so it is enabled as a non-default option.

Fixes #2063