cvar: ignore the class field in CVARs
Many of our cvars were unnecessarily set in the "device" class,
probably because they were copy-pasted from other places.  The device
class is only meant for CVARs where the CVAR itself can be overridden
by the MPID_ version of the same CVAR.  This doesn't apply to any of
the CVARs modified in this patch.  It is unclear if it applies to any
CVARs at all.