ch4: refactor collective selection code
The code was overly complicated based on the assumption that the ch4
layer would handle the selection logic for all layers.  This is ugly
and breaks abstractions.  Instead, we want to move to a model where
each layer does its own tuning (mostly) independently.

This patch significantly simplifies the container management code.  In
fact, the ch4-level containers simply select the composition algorithm
and nothing else.