coll: move device coll init hook outside of the MPI-level coll init
This is the first step in making the coll init hook to a general post
commit hook for communicators.

CH3 used the coll init hook in a hacky way for dynamic processes.
While this change should technically affect it, its coll init hook is
a no-op.  So it should not matter.