config: enhance PAC_SET_HEADER_LIB_PATH
Remove the default option in --with-$1=[[PATH]]. We don't use default in
our current scripts, and it is more confusing when we use it especially
because the actual macros are hidden inside aclocal_libs.m4. If we need
to, it is cleaner to manage it explicitly.

Add case to bypass setting flags if user set "system" or "embedded" or
"no".