config: enhance PAC_SET_HEADER_LIB_PATH
Remove the default option in PAC_SET_HEADER_LIB_PATH. We don't use default in
our current scripts, and it is more confusing when we use it especially
because the actual macros are hidden inside aclocal_libs.m4. If we need
to, it is cleaner to manage it explicitly.

We also tune the semantics of `--with-[libname]` option a little. When
user explicitly pass this option, naturally user want to use this
library, so we should throw error (to be enforced in next commit) if the
library is not found. On the other hand, `--without-[libname]` should
have the same effect as the library not found. We also allow "embedded"
option, which use the embedded library, and throw error if embedded
version doesn't exist.

As an example, simply `--with-hwloc` means require hwloc and the hwloc
library should be found in the system/current CPATH and LIBRARY_PATH.
`--with-hwloc=path` will add the paths to CPPFLAGS and LDFLAGS.
`--without-hwloc` should throw an error -- since we require hwloc --
even if hwloc can be found. `--with-hwloc=embedded` will use the
embedded hwloc, and we'll try keep the hwloc symbols hidden if possible.

If user didn't supply `--with[out]-hwloc` option, assuming we don't have
embedded hwloc, we'll check for hwloc from system path. Assuming
hwloc is optional, then we'll either enable or disable features that
related to hwloc depend on the configure check results.