config: remove the override of have_hwloc on cygwin
This hack is based on very old version of hwloc and probably an old
version of cygwin. It may as well still be true -- that hwloc does not
build on cygwin -- but we should add them back once we confirm rather
than leaving obsolete comments.

A proper way to disable hwloc on certain system is to use
--with-hwloc=no.