config: externally configure hwloc
Hwloc can be subconfigured into a convenience library with no issue, so
do that.

PAC_CHECK_HEADER_LIB_OPTIONAL should take care of all cases when user
pass --with-hwloc explicitly. We only need choose the embedded as
default and sub-configure it accordingly.

Remove the override of have_hwloc on cygwin. --without-hwloc should
serve the purpose.