ch4/ucx: remove OPA_read_write_barrier
The full memory barrier is added on the concern that another progress
thread may read the status before its completion. This commit removes it
on the reason that we already use atomic completion counter and status
shouldn't be read before cc reaches 0. So if there is pre-mature access
to status, it should be fixed at the site of access (for not waiting on
the cc in the first place).

Besides, this is not unique to ucx. So having a special solution for ucx
doesn't seem right anyway.