init_shm: remove code using inter-process lock based atomics
MPL atomics has not implemented inter-process lock based atomics yet.
Since it is rare for compilers not to support atomics natively now, we
simply raise error if it is not available. Usually when we fall-back to
lock-based atomics, it is some glitch in the configuration that need be
fixed instead.