ch4: replace cacheline-alignment with explicit union
Enforce minimul struct size using align attribute (to break false
sharing) is nice when it works. But it ultimately is a compiler
extension and may not always work. When it doesn't work -- e.g. when
`--enable-thread=single --with-thread-package=none` -- it is difficult
to understand why and how to fix. This patch replace it with explict
union for more reliability.