ch4: Expand the addressing in OFI to use multiple NICs
Expands the OFI addressing collection and exchange to capture
information from multiple NICs and insert it into the CH4-level address
vector as well as the OFI address vectors attached to each endpoint.

The OFI-level context field (MPIDI_OFI_global.ctx) is used to keep track
of the various OFI-isms needed to conduct communication on separate nics
(e.g. domains, endpoints, address vectors, etc.). This patch accounts
for both the domain-centric and endpoint-cetric methods of handling
multiple OFI vnis.

This patch assumes that in the future, another patch will be added which
will detect and open as many providers as there are nics. In the
meantime, context 0 is still used for all communication.

For UCX, the only change is to be compatible with the changes in the BC
code.