test/dtpools: remove lower bound for buffer size
Currently the object create code in dtpools only allows maxbufsize that
is bigger than a certain arbitrary size (100000). This is to avoid that
the generated buffer size for objects is always bigger than the allowed
max buffer size. In this cases dtpools tries creating the object again
and might never succeed ending up in an infinite loop. Make the code
break out of the loop after a max attempt threshold is reached.