thread: define MPID_THREAD_YIELD to allow choice of yield
The ch3 design has reasonable fairness in progress lock yielding while
being very sensitive to the cost in thread yield, probably becasue only
one thread is making real progress. Therefore it requires the yield as
light as possible.

The ch4 design progress is heavy, and once yield, the other threads may
pick up the progress equivalently. It is detrimental to have a lock
monopoly. Therefore, it requires the yield to be fair while it can
afford cost a bit more.