ch4/ofi: Update data structures to support striping
This coomit adds new fields to MPIDI_OFI_global_t, MPIDI_OFI_send_control_t,
and MPIDI_OFI_huge_recv_t for large message striping across multiple NICs.
In MPIDI_OFI_global_t a new variable is added to keep track of the striping
threshold. When message sizes exceed this threshold, they will be striped across
multiple NICs. In MPIDI_OFI_send_control_t, an array of rma keys are stored
instead of a single key, because for striping the buffer needs to be registered
in all the NICs (instead of just one). In MPIDI_OFI_huge_recv_t, two new fields
are added, stripe_size (to store the stripe size for a message) and bytes_recvd
(to keep track of the amount of data received so far).