test/rma: assert on memory allocation failures
In our jenkins testing, we are failing some of the RMA tests on
solaris with an unrecognizable error.  The main problem seems to be
that we are treating memory allocation failures as a "soft error".
Without the allocated memory, we cannot proceed further.  This patch
converts those to an assert.