romio: squash unused function warning
The MPIR_Ext_mutex_finalize is not being called. It is generally not a
problem since we use static mutex in posix and does not really destroy.
For dynamic mutex such as in Argobots, skipping the finalize will leak a
small amount of memory. However, we do not have a place to call the
finalization currently. If will require a ROMIO interface change so that
users will always init/finalize ROMIO explicitly. Until then, this is
best we can do.