typerep: move dataloop creation to typerep
Typerep was assuming that the datatype engine (dataloop in this case)
has knowledge about the internals of MPICH and that it can infer the
actual construction of the datatype from the MPI datatype object.
This patch undoes this assumption and moves the MPI datatype parsing
code to typerep.  The datatype engine itself does not need to be aware
of this information.