typerep: added a new datatype engine for yaksa
Some notes:

1. This is a new datatype engine behind the typerep interface, that
mostly redirects all typerep calls to the yaksa API.  For the cases
where yaksa doesn't provide direct support, the typerep glue layer
adds the necessary functionality.

2. This new backend is disabled by default.  Setting
--with-datatype-engine=auto would enable it for ch4, but not for ch3.
That's because ch3 has a number of assumptions related to how bytes
are packed.  In particular, ch3 assumes that basic types can be split
while packing.  yaksa does not allow splitting of basic types in order
to maintain their atomicity, for example, during an accumulate
operation.  Users can force the usage of yaksa for all devices by
setting --with-datatype-engine=yaksa.

3. MPICH calculates and caches some information (such as size, extent,
number of contiguous elements, etc.) in its MPIR_Datatype structure.
Because we go through Typerep, there is an underlying assumption that
this metadata matches the metadata that yaksa detects.  This seems to
be true today, but there is no guarantee that it will continue to be
true.  The right thing to do is for mpich to not recalculate this
metadata, but to query it from the datatype engine.  But dataloop does
not seem to have this capability, so that's a nonstarter till we
delete dataloop.