comm: commenting out the MPIR_Allreduce in sched_cb_gcn_allocate_cid
Running an MPIR_Allreduce inside a progress not only invokes a
collective operation of arbitary algorithm, it introduces nested
progress. Since all it does here is to simply getting a slightly better
error message, it appears not to worth the complications.

I am simply commenting off the code here and leaving a FIXME to resolve
later. We are already inside a non-blocking collective operation, so it
appears that we should have ways to propagate the errors rather than
launching another collective.