ch4/ofi: simplify send/recv for iov messages
The IOV based send/recv functions are only called when the total
message size is smaller than the maximum allowed bytes.  So the
additional functionality in these functions to check if each of the
IOV segments is less than the maximum allowed bytes in unnecessary.
In fact, that functionality seems incorrect because the maximum
allowed bytes is a limit for the entire message, and not for each IOV
segment.