mpl: remove windows specific types
This change was left over from the time when we removed windows
support in mpich.

This abstraction disallows us from using iovecs created within MPICH
directly in other external libraries (such as yaksa or OFI).  So we'd
need to convert the IOVEC from one format to another each time, which
is very expensive.  "struct iovec" is a part of POSIX 2001, which we
assume in MPICH anyway.