tests: bug fix in pairtype test
The MPI standard does not allow us to receive a part of a basic
datatype in a message.  This test was buggy and was expecting the MPI
library to be able to get a subset of a pairtype datatype.  With
dataloop, this test is able to pass because dataloop converts
pairtypes to structs.  But this will not pass with yaksa, where
pairtypes are builtin.