logging: cleanup function debug state information
We have not been consistent with updating debug macros while updating
function names.  This is a mostly script-generated cleanup of the
debug macros, although it is hard to keep it clean consistently.  We
should consider an automated way to generate this in the future, or
perhaps rely on __func__ like macros, instead.

ch3 is left alone to avoid code churn.