datatype: ensure element is not resized before coalescing
In MPIR_Dataloop_create_contiguous, we optimize when the lower layer
type is also a contiguous type, but we need ensure the element is not
resized. This happens when we have Contig/Resized/Resized/Basic, when
the first resized is turned into contig, but the second resized is
duped resulting in discrepancy in extent/size get from type and the
actual element.