group: make group routines thread-safe
The current code use a flag field within a group structure to faciliate
group calculations, which made it thread unsafe if mutliple threads are
operating on the same group. Since we only need a temporary array for
these operations, we can simply malloc the temporary flags array and
make all these routines thread safe. These routines are not in a
performance critical path. Even considering performance, the overhead of
adding a critical section will negate the cost of malloc/free.

NOTE: they only become an issue now as we move to per-vci thread
granularity and removing the global locks.