coll: add MPIR_ERR_COLL_CHECKANDCONT macro
We repeat the multi-line error checking in collectives many times and it
has become difficult to read. In addition, we may want to tune the error
propagation mechanism and having them scattered make tuning very
difficult. This commit adds a macro to achieve better readability, and a
single place for tuning the behavior.