ch4/ofi: factor out the flush_send_queue
This allreduce function in ofi finalize has baffled me for quite a
while. It routinely breaks because the allreduce collective may choose
some algorithm that rely on components that are unavilable during
finalize. And its purpose were quite misleading. On one hand, it uses
variable name such as barrier, but its purpose has nothing to do with
barrier. In fact, its purpose is to enqueue a send message then
explicitly wait until the send queue empties, before moving on freeing
everything. It is merely using the side effects of allreduce that every
process need place some outgoing messages and there is an explicit wait
in the collective.

However, this solution is hacky. Not only it invokes more funcationality
than what is needed, placing restriction on different parts of code
during finalize, but it is also unreliable. For example, there is no
control that some of the outgoing messages may be sent via shmem,
defeating its purpose of flushing ofi send queue. In addition, when we
enable multiple vci, we will need flush send queue in every vci channel.
A single allreduce is insufficient.

This commit factors the function out for later update.