ch3: move completed callback to MPIR_Request_free_hook
In ch3, each request can have a request_completed_cb to be called at
completion, currently from MPID_Request_complete. This is used for
nonblocking requests that we don't want to explicitly wait for but have
them call back on completion instead. It is only used in two places, in
MPID_Win_fence calling an MPIR_Ibarrier and in MPID_Win_start calling a
MPID_Irecv. Now that non-blocking collective request are completed by
MPIR_Request_complete instead, we need move the callback to
MPID_Request_free_hook for the callback mechanism to still work for all
requests in ch3.