config: cleanup config options for locality
Because we already have MPIR_CVAR_ODD_EVEN_CLIQUES,
--enable-dbg-localoddeven is redudant and rarely useful, therefore,
removed.

We'll use --enable-nolocal as an option to statically disable shared
memory path (not just locality). For locality, it will have the same
effect as before, that is, no locality. To better reflect its usage
(especially, no longer just a debug option) we name it option and help
string more approprately.