mpl: tell the compiler that the size is within bounds
GCC seems to have an overly aggressive check to see the maximum bounds
of each variable.  When an integer is passed to the allocation, in
theory, it could have a negative value (which, when cast to size_t
becomes a very large value).  GCC complains about this case with
[-Walloc-size-larger-than=].  This patch converts the compile-time
check to a runtime check.

This is somewhat of a hack fix, because the correct fix would be to
change the calling site so that there is no implicit cast from a
signed integer to size_t.  But that would create a lot of code churn,
which this patch is trying to avoid.